+++ personal-stock-tracking/jest.config.mjs
@@ -1,8 +1,0 @@
-import nextJest from 'next/jest.js';
-
-const createJestConfig = nextJest({ dir: './' });
-
-export default createJestConfig({
-  testEnvironment: 'jest-environment-jsdom',
-  setupFilesAfterEnv: ['<rootDir>/src/test/setup.ts']
-});